More Linux 64 bit fixes

+++ etmain/src/cgame/cg_main.c
@@ -55,7 +55,7 @@ This must be the very first function compiled into the .q3vm file
 #if __GNUC__ >= 4
 #pragma GCC visibility push(default)
 #endif
-int vmMain(int command, int arg0, int arg1, int arg2, int arg3, int arg4, int arg5, int arg6, int arg7, int arg8, int arg9,
+intptr_t vmMain(int command, int arg0, int arg1, int arg2, int arg3, int arg4, int arg5, int arg6, int arg7, int arg8, int arg9,
 		   int arg10, int arg11)
 {
 #if __GNUC__ >= 4
